layers: Create CoreCheck CommandBuffer validation file
