+++ BUILD.gn
@@ -114,6 +114,7 @@ core_validation_sources = [
   "layers/core_checks/android_validation.cpp",
   "layers/core_checks/buffer_validation.cpp",
   "layers/core_checks/cmd_buffer_dynamic_validation.cpp",
+  "layers/core_checks/cmd_buffer_validation.cpp",
   "layers/core_checks/copy_blit_resolve_validation.cpp",
   "layers/core_checks/core_validation.cpp",
   "layers/core_checks/core_validation.h",
